requirements.txt
streamlit==1.32.2
python-docx==1.1.0
python-dotenv
requests
streamlit-lottie

